Fix issue with sampleFromDistribution
In the GravesLSTMCharModellingExample class, the sampleFromDistribution method was assuming that its input distribution would sum to 1. However, rounding error means that this is not guaranteed to be the case. I had the demo crash because the randomly chosen number was 0.99999993, but the sum was only 0.99999985.

Fix this issue by having sampleFromDistribution try up to 10 times.