Properly truncate short reviews during prediction (fixes #761)
When training, we learn `maxLength`, which is the largest number of tokens found in a single review.

When predicting, we initialize a feature vector with the same size as the document to be predicted *or* `maxLength`, whatever is *smaller*. This avoids filling the feature vector with invalid data in the case of a too short document, thus leading to a wrong prediction. This also avoids initializing a feature vector which is longer than the longest training sample.

TEST=Run the main method in `Word2VecSentimentRNN`. `p(negative)` should be greater than `p(positive)` for document `neg/12100_1.txt`.

Co-authored-by: Markus Kreuzthaler <markus.kreuzthaler@gmail.com>